Fix account creation
