+++ services/ai-bot/pod-ai-bot/src/controller.ts
@@ -14,16 +14,20 @@
 //
 
 import { Db, Collection } from 'mongodb'
-import { getWorkspaceId, MeasureContext, toWorkspaceString, WorkspaceId } from '@hcengineering/core'
+import { getWorkspaceId, MeasureContext, systemAccountEmail, toWorkspaceString, WorkspaceId } from '@hcengineering/core'
 import { aiBotAccountEmail, AIBotTransferEvent } from '@hcengineering/ai-bot'
 import { WorkspaceInfoRecord } from '@hcengineering/server-ai-bot'
 import { getTransactorEndpoint } from '@hcengineering/server-client'
 import { generateToken } from '@hcengineering/server-token'
+import { WorkspaceLoginInfo } from '@hcengineering/account'
 
 import { WorkspaceClient } from './workspaceClient'
+import { assignBotToWorkspace, getWorkspaceInfo } from './account'
 
 const POLLING_INTERVAL_MS = 5 * 1000 // 5 seconds
 const CLOSE_INTERVAL_MS = 10 * 60 * 1000 // 10 minutes
+const ASSIGN_WORKSPACE_DELAY_MS = 5 * 1000 // 5 secs
+const MAX_ASSIGN_ATTEMPTS = 5
 
 export class AIBotController {
   private readonly workspaces: Map<string, WorkspaceClient> = new Map<string, WorkspaceClient>()
@@ -34,6 +38,9 @@ export class AIBotController {
   private readonly workspacesInfoCollection: Collection<WorkspaceInfoRecord>
 
   private readonly intervalId: NodeJS.Timeout
+
+  assignTimeout: NodeJS.Timeout | undefined
+  assignAttempts = 0
 
   constructor (mongoDb: Db, ctx: MeasureContext) {
     this.db = mongoDb
@@ -84,11 +91,65 @@ export class AIBotController {
     }
   }
 
+  private async getWorkspaceInfo (ws: WorkspaceId): Promise<WorkspaceLoginInfo | undefined> {
+    const systemToken = generateToken(systemAccountEmail, ws)
+    for (let i = 0; i < 5; i++) {
+      try {
+        const info = await getWorkspaceInfo(systemToken)
+
+        if (info == null) {
+          this.ctx.warn('Cannot find workspace info', ws)
+          await wait(ASSIGN_WORKSPACE_DELAY_MS)
+          continue
+        }
+
+        return info
+      } catch (e) {
+        this.ctx.error('Error during get workspace info:', { e })
+        await wait(ASSIGN_WORKSPACE_DELAY_MS)
+      }
+    }
+  }
+
+  private async assignToWorkspace (ws: WorkspaceId): Promise<void> {
+    clearTimeout(this.assignTimeout)
+    try {
+      const info = await this.getWorkspaceInfo(ws)
+
+      if (info === undefined) {
+        void this.closeWorkspaceClient(ws)
+        return
+      }
+
+      if (info.creating === true) {
+        this.ctx.info('Workspace is creating -> waiting...', ws)
+        this.assignTimeout = setTimeout(() => {
+          void this.assignToWorkspace(ws)
+        }, ASSIGN_WORKSPACE_DELAY_MS)
+        return
+      }
+
+      const result = await assignBotToWorkspace(ws)
+      this.ctx.info('Assign to workspace result: ', { result, workspace: ws.name })
+    } catch (e) {
+      this.ctx.error('Error during assign workspace:', { e })
+      if (this.assignAttempts < MAX_ASSIGN_ATTEMPTS) {
+        this.assignAttempts++
+        this.assignTimeout = setTimeout(() => {
+          void this.assignToWorkspace(ws)
+        }, ASSIGN_WORKSPACE_DELAY_MS)
+      } else {
+        void this.closeWorkspaceClient(ws)
+      }
+    }
+  }
+
   async initWorkspaceClient (workspaceId: WorkspaceId, info: WorkspaceInfoRecord): Promise<void> {
     const workspace = toWorkspaceString(workspaceId)
 
     if (!this.workspaces.has(workspace)) {
       this.ctx.info('Listen workspace: ', { workspace })
+      await this.assignToWorkspace(workspaceId)
       const token = generateToken(aiBotAccountEmail, workspaceId)
       const endpoint = await getTransactorEndpoint(token)
       this.workspaces.set(
@@ -144,4 +205,12 @@ export class AIBotController {
       { $set: { avatarPath: path, avatarLastModified: lastModified } }
     )
   }
+}
+
+async function wait (delay: number): Promise<void> {
+  await new Promise<void>((resolve) => {
+    setTimeout(() => {
+      resolve()
+    }, delay)
+  })
 }
